requirements.txt
streamlit
llama-cpp-python
langchain
chromadb
sentence-transformers
bs4
requests
